requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
PyJWT==2.7.0
qrcode==7.4.2
cryptography==39.0.2
bleach>=6.0.0,<7.0.0
marshmallow==3.19.0
gunicorn==20.1.0

# New dependencies for comprehensive features
requests==2.31.0
razorpay==1.3.0
googlemaps==4.10.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Pillow==9.5.0
